# swagger_server/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: Invoice Management
  version: 0.0.1
servers:
- url: /
...
paths:
  /api/v{version}/auth:
    post:
      tags:
      - Authentication
      operationId: api_vversion_auth_post
      parameters:
      - name: version
        in: path
        description: Version number
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        description: Request to authenticate
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Login'
      responses:
        "201":
          description: Successfully authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      x-openapi-router-controller: swagger_server.controllers.authentication_controller
  /api/v{version}/users/sso:
    post:
      tags:
      - Authentication
      operationId: api_vversion_users_sso_post
      parameters:
      - name: version
        in: path
        description: Version number
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        description: User authentication by SSO
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Token'
      responses:
        "201":
          description: User successfully registered by SSO
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      x-openapi-router-controller: swagger_server.controllers.authentication_controller
...
  /api/v{version}/categories:
    get:
      tags:
      - Category
      operationId: api_vversion_categories_get
      parameters:
      - name: version
        in: path
        description: Version number
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: name
        in: query
        description: Category's name
        required: false
        style: form
        explode: true
        schema:
          type: string
      responses:
        "200":
          description: Categories founded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Categories'
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
      - user_auth: []
      x-openapi-router-controller: swagger_server.controllers.category_controller
    post:
      tags:
      - Category
      operationId: api_vversion_categories_post
      parameters:
      - name: version
        in: path
        description: Version number
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        description: Register category
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Category'
      responses:
        "201":
          description: Expense's ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponse'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
      - user_auth: []
      x-openapi-router-controller: swagger_server.controllers.category_controller
  /api/v{version}/categories/{categoryID}:
    get:
      tags:
      - Category
      operationId: api_vversion_categories_category_idget
      parameters:
      - name: version
        in: path
        description: Version number
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: categoryID
        in: path
        description: Category's ID
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: Category founded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
      - user_auth: []
      x-openapi-router-controller: swagger_server.controllers.category_controller
    put:
      tags:
      - Category
      operationId: api_vversion_categories_category_idput
      parameters:
      - name: version
        in: path
        description: Version number
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: categoryID
        in: path
        description: Category's ID
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int32
      requestBody:
        description: Update category
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Category'
      responses:
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
      - user_auth: []
      x-openapi-router-controller: swagger_server.controllers.category_controller
  /api/v{version}/categories/{categoryID}/archives:
    put:
      tags:
      - Category
      operationId: api_vversion_categories_category_id_archives_put
      parameters:
      - name: version
        in: path
        description: Version number
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: categoryID
        in: path
        description: Category's ID
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int32
      responses:
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
      - user_auth: []
      x-openapi-router-controller: swagger_server.controllers.category_controller
...
components:
  schemas:
    Login:
      type: object
      properties:
        login:
          type: string
        password:
          type: string
    LoginResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: '#/components/schemas/User'
      example:
        user:
          password: password
          company_name: company_name
          name: name
          phone_number: phone_number
          cnpj: cnpj
          email: email
        token: token
    Token:
      type: object
      properties:
        login:
          type: string
        app_token:
          type: string
    User:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        cnpj:
          type: string
        company_name:
          type: string
        phone_number:
          type: string
      example:
        password: password
        company_name: company_name
        name: name
        phone_number: phone_number
        cnpj: cnpj
        email: email
...
    Category:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
      example:
        name: name
        description: description
    CategoryResponse:
      type: object
      properties:
        category_id:
          type: integer
      example:
        category_id: 0
    Categories:
      type: object
      properties:
        count:
          type: integer
        categories:
          type: array
          items:
            $ref: '#/components/schemas/Category'
      example:
        count: 0
        categories:
        - name: name
          description: description
        - name: name
          description: description
...
    Error:
      type: object
      properties:
        error:
          type: string
  securitySchemes:
    user_auth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      x-bearerInfoFunc: swagger_server.controllers.authorization_controller.check_user_auth